cf-sync-api-specs/requirements.txt
google-cloud-storage
pinecone>=7.3.0
google-genai
requests
pyyaml
